+++ server/src/test/java/org/opensearch/action/search/AbstractSearchAsyncActionTests.java
@@ -338,7 +338,14 @@ public class AbstractSearchAsyncActionTests extends OpenSearchTestCase {
         SearchQueryThenFetchAsyncAction action = createSearchQueryThenFetchAsyncAction(requestOperationListeners);
         action.start();
         assertEquals(1, testListener.getPhaseCurrent(action.getSearchPhaseName()));
-        action.onPhaseFailure(new SearchPhase("test") {
+        action.onPhaseFailure(new SearchPhase("none") {
+            @Override
+            public void run() {
+
+            }
+        }, "message", null);
+        assertEquals(1, testListener.getPhaseCurrent(action.getSearchPhaseName()));
+        action.onPhaseFailure(new SearchPhase(action.getName()) {
             @Override
             public void run() {
 
@@ -352,14 +359,14 @@ public class AbstractSearchAsyncActionTests extends OpenSearchTestCase {
         );
         searchDfsQueryThenFetchAsyncAction.start();
         assertEquals(1, testListener.getPhaseCurrent(searchDfsQueryThenFetchAsyncAction.getSearchPhaseName()));
-        searchDfsQueryThenFetchAsyncAction.onPhaseFailure(new SearchPhase("test") {
+        searchDfsQueryThenFetchAsyncAction.onPhaseFailure(new SearchPhase(searchDfsQueryThenFetchAsyncAction.getName()) {
             @Override
             public void run() {
 
             }
         }, "message", null);
-        assertEquals(0, testListener.getPhaseCurrent(action.getSearchPhaseName()));
-        assertEquals(0, testListener.getPhaseTotal(action.getSearchPhaseName()));
+        assertEquals(0, testListener.getPhaseCurrent(searchDfsQueryThenFetchAsyncAction.getSearchPhaseName()));
+        assertEquals(0, testListener.getPhaseTotal(searchDfsQueryThenFetchAsyncAction.getSearchPhaseName()));
 
         FetchSearchPhase fetchPhase = createFetchSearchPhase();
         ShardId shardId = new ShardId(randomAlphaOfLengthBetween(5, 10), randomAlphaOfLength(10), randomInt());
@@ -368,7 +375,7 @@ public class AbstractSearchAsyncActionTests extends OpenSearchTestCase {
         action.skipShard(searchShardIterator);
         action.executeNextPhase(action, fetchPhase);
         assertEquals(1, testListener.getPhaseCurrent(fetchPhase.getSearchPhaseName()));
-        action.onPhaseFailure(new SearchPhase("test") {
+        action.onPhaseFailure(new SearchPhase(fetchPhase.getName()) {
             @Override
             public void run() {
 
@@ -401,6 +408,30 @@ public class AbstractSearchAsyncActionTests extends OpenSearchTestCase {
         assertEquals(0, searchPhaseExecutionException.getSuppressed().length);
         assertEquals(nodeLookups, resolvedNodes);
         assertEquals(requestIds, releasedContexts);
+    }
+
+    public void testOnPhaseStart() {
+        ClusterSettings clusterSettings = new ClusterSettings(Settings.EMPTY, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS);
+        SearchRequestStats testListener = new SearchRequestStats(clusterSettings);
+
+        final List<SearchRequestOperationsListener> requestOperationListeners = new ArrayList<>(List.of(testListener));
+        SearchQueryThenFetchAsyncAction action = createSearchQueryThenFetchAsyncAction(requestOperationListeners);
+
+        action.onPhaseStart(new SearchPhase("test") {
+            @Override
+            public void run() {}
+        });
+        action.onPhaseStart(new SearchPhase("none") {
+            @Override
+            public void run() {}
+        });
+        assertEquals(0, testListener.getPhaseCurrent(action.getSearchPhaseName()));
+
+        action.onPhaseStart(new SearchPhase(action.getName()) {
+            @Override
+            public void run() {}
+        });
+        assertEquals(1, testListener.getPhaseCurrent(action.getSearchPhaseName()));
     }
 
     public void testShardNotAvailableWithDisallowPartialFailures() {